chore: move combat modules
